Reformatting.

+++ src/fheroes2/agg/mus.cpp
@@ -33,83 +33,93 @@ namespace MUS
 {
     const struct
     {
-	int type;
-	const char* string;
+        int type;
+        const char *string;
     } musmap[] = {
-        { UNUSED,       ""                    },
-        { DATATRACK,    ""                    },
-        { BATTLE1,      "Battle (1)"          },
-        { BATTLE2,      "Battle (2)"          },
-        { BATTLE3,      "Battle (3)"          },
-        { BARBARIAN,    "Barbarian Castle"    },
-        { SORCERESS,    "Sorceress Castle"    },
-        { WARLOCK,      "Warlock Castle"      },
-        { WIZARD,       "Wizard Castle"       },
-        { NECROMANCER,  "Necromancer Castle"  },
-        { KNIGHT,       "Knight Castle"       },
-        { LAVA,         "Lava Theme"          },
-        { WASTELAND,    "Wasteland Theme"     },
-        { DESERT,       "Desert Theme"        },
-        { SNOW,         "Snow Theme"          },
-        { SWAMP,        "Swamp Theme"         },
-        { BEACH,        "Ocean Theme"         },
-        { DIRT,         "Dirt Theme"          },
-        { GRASS,        "Grass Theme"         },
-        { LOSTGAME,     "Lost Game"           },
-        { WEEK1,        "Week (1)"            },
-        { WEEK2_MONTH1, "Week (2) Month (1)"  },
-        { MONTH2,       "Month (2)"           },
-        { PUZZLE,       "Map Puzzle"          },
-        { ROLAND,       "Roland's Campaign"   },
-        { CARAVANS,     "25"                  },
-        { CARAVANS_2,   "26"                  },
-        { CARAVANS_3,   "27"                  },
-        { COMPUTER,     "28"                  },
-        { BATTLEWIN,    "29"                  },
-        { BATTLELOSE,   "30"                  },
-        { DEATH,        "31"                  },
-        { WATERSPRING,  "32"                  },
-        { ARABIAN,      "33"                  },
-        { NOMADTENTS,   "34"                  },
-        { TREEHOUSE,    "35"                  },
-        { DEMONCAVE,    "36"                  },
-        { EXPERIENCE,   "37"                  },
-        { SKILL,        "38"                  },
-        { WATCHTOWER,   "39"                  },
-        { EVENT15,      "40"                  },
-        { NEWS,         "41"                  },
-        { MAINMENU,     "Main Menu"           },
-        { VICTORY,      "Scenario Victory"    },
-        { UNKNOWN,      "UNKNOWN"             }
+            {UNUSED,       ""},
+            {DATATRACK,    ""},
+            {BATTLE1,      "Battle (1)"},
+            {BATTLE2,      "Battle (2)"},
+            {BATTLE3,      "Battle (3)"},
+            {BARBARIAN,    "Barbarian Castle"},
+            {SORCERESS,    "Sorceress Castle"},
+            {WARLOCK,      "Warlock Castle"},
+            {WIZARD,       "Wizard Castle"},
+            {NECROMANCER,  "Necromancer Castle"},
+            {KNIGHT,       "Knight Castle"},
+            {LAVA,         "Lava Theme"},
+            {WASTELAND,    "Wasteland Theme"},
+            {DESERT,       "Desert Theme"},
+            {SNOW,         "Snow Theme"},
+            {SWAMP,        "Swamp Theme"},
+            {BEACH,        "Ocean Theme"},
+            {DIRT,         "Dirt Theme"},
+            {GRASS,        "Grass Theme"},
+            {LOSTGAME,     "Lost Game"},
+            {WEEK1,        "Week (1)"},
+            {WEEK2_MONTH1, "Week (2) Month (1)"},
+            {MONTH2,       "Month (2)"},
+            {PUZZLE,       "Map Puzzle"},
+            {ROLAND,       "Roland's Campaign"},
+            {CARAVANS,     "25"},
+            {CARAVANS_2,   "26"},
+            {CARAVANS_3,   "27"},
+            {COMPUTER,     "28"},
+            {BATTLEWIN,    "29"},
+            {BATTLELOSE,   "30"},
+            {DEATH,        "31"},
+            {WATERSPRING,  "32"},
+            {ARABIAN,      "33"},
+            {NOMADTENTS,   "34"},
+            {TREEHOUSE,    "35"},
+            {DEMONCAVE,    "36"},
+            {EXPERIENCE,   "37"},
+            {SKILL,        "38"},
+            {WATCHTOWER,   "39"},
+            {EVENT15,      "40"},
+            {NEWS,         "41"},
+            {MAINMENU,     "Main Menu"},
+            {VICTORY,      "Scenario Victory"},
+            {UNKNOWN,      "UNKNOWN"}
     };
-    
+
     const std::string GetString(int mus, bool shortname)
     {
-      std::stringstream sstream;
-      sstream << std::setw(2) << std::setfill('0') << (int)mus;
-      if(shortname)
-      sstream << ".ogg";
-      else
-      sstream << " " <<
-	(UNUSED <= mus && UNKNOWN > mus ? musmap[mus].string : musmap[UNKNOWN].string) << ".ogg";
-      return sstream.str();
+        std::stringstream sstream;
+        sstream << std::setw(2) << std::setfill('0') << mus;
+        if (shortname)
+            sstream << ".ogg";
+        else
+            sstream << " " <<
+                    (UNUSED <= mus && UNKNOWN > mus ? musmap[mus].string : musmap[UNKNOWN].string) << ".ogg";
+        return sstream.str();
     }
 }
 
 int MUS::FromGround(int ground)
 {
-    switch(ground)
+    switch (ground)
     {
-        case Maps::Ground::DESERT:	return DESERT;
-        case Maps::Ground::SNOW:	return SNOW;
-        case Maps::Ground::SWAMP:	return SWAMP;
-        case Maps::Ground::WASTELAND:	return WASTELAND;
-        case Maps::Ground::BEACH:	return BEACH;
-        case Maps::Ground::LAVA:	return LAVA;
-        case Maps::Ground::DIRT:	return DIRT;
-        case Maps::Ground::GRASS:	return GRASS;
-        case Maps::Ground::WATER:	return BEACH;
-        default: break;
+        case Maps::Ground::DESERT:
+            return DESERT;
+        case Maps::Ground::SNOW:
+            return SNOW;
+        case Maps::Ground::SWAMP:
+            return SWAMP;
+        case Maps::Ground::WASTELAND:
+            return WASTELAND;
+        case Maps::Ground::BEACH:
+            return BEACH;
+        case Maps::Ground::LAVA:
+            return LAVA;
+        case Maps::Ground::DIRT:
+            return DIRT;
+        case Maps::Ground::GRASS:
+            return GRASS;
+        case Maps::Ground::WATER:
+            return BEACH;
+        default:
+            break;
     }
 
     return UNKNOWN;
@@ -117,15 +127,22 @@ int MUS::FromGround(int ground)
 
 int MUS::FromRace(int race)
 {
-    switch(race)
+    switch (race)
     {
-        case Race::KNGT:	return KNIGHT;
-        case Race::BARB:	return BARBARIAN;
-        case Race::SORC:	return SORCERESS;
-        case Race::WRLK:	return WARLOCK;
-        case Race::WZRD:	return WIZARD;
-        case Race::NECR:	return NECROMANCER;
-        default: break;
+        case Race::KNGT:
+            return KNIGHT;
+        case Race::BARB:
+            return BARBARIAN;
+        case Race::SORC:
+            return SORCERESS;
+        case Race::WRLK:
+            return WARLOCK;
+        case Race::WZRD:
+            return WIZARD;
+        case Race::NECR:
+            return NECROMANCER;
+        default:
+            break;
     }
 
     return UNKNOWN;
@@ -133,10 +150,10 @@ int MUS::FromRace(int race)
 
 int MUS::FromMapObject(int object)
 {
-    if(Settings::Get().MusicMIDI())
+    if (Settings::Get().MusicMIDI())
         return MUS::UNKNOWN;
-    
-    switch(object)
+
+    switch (object)
     {
         case MP2::OBJ_WITCHSHUT:
         case MP2::OBJ_FORT:
@@ -170,7 +187,7 @@ int MUS::FromMapObject(int object)
 
         case MP2::OBJ_EVENT:
             return MUS::NEWS;
-            
+
         default:
             return MUS::UNKNOWN;
     }
@@ -178,12 +195,16 @@ int MUS::FromMapObject(int object)
 
 int MUS::GetBattleRandom(void)
 {
-    switch(Rand::Get(1, 3))
+    switch (Rand::Get(1, 3))
     {
-	case 1:	return BATTLE1;
-	case 2:	return BATTLE2;
-	case 3:	return BATTLE3;
-	default: break;
+        case 1:
+            return BATTLE1;
+        case 2:
+            return BATTLE2;
+        case 3:
+            return BATTLE3;
+        default:
+            break;
     }
     return UNKNOWN;
 }
